third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-018-expected.txt


FAIL CSS Test: @viewport constrained - Small non-auto height not affecting width. assert_equals: expected 1 but got 600
